refactor db continued
